fix volume

+++ esphome/components/es8311/es8311.cpp
@@ -182,7 +182,7 @@ void ES8311Component::dump_config() {
 
 void ES8311Component::set_volume(float volume) {
   volume = clamp(volume, 0.0f, 1.0f);
-  uint8_t reg32 = remap<float, uint8_t>(volume, 0.0f, 1.0f, 0, 255);
+  uint8_t reg32 = remap<uint8_t, float>(volume, 0.0f, 1.0f, 0, 255);
   ES8311_WRITE_BYTE(ES8311_REG32_DAC, reg32);
 }
